+++ src/Model/WslElement.psm1
@@ -1,0 +1,12 @@
+
+Class WslElement
+{
+
+    [string]$name
+    [bool]$default=$false
+    [bool]$running=$false
+    [int]$wslVersion=1
+    [string]$from="-"
+    [string]$creation="-"
+
+}